Increase cache duration to 60 minutes and adjust spacing in exchange hides and elixirs widget layout; implement CalculateMaxProfit class for profit calculations.
